fixing the controllers

+++ controllers/campgrounds.js
@@ -1,15 +1,15 @@
-const Campground = require('../models/campground')
-const mbxGeocoding = require('@mapbox/mapbox-sdk/services/geocoding')
-const mapBoxToken = process.env.MAPBOX_TOKEN
-const geocoder = mbxGeocoding({accessToken:mapBoxToken})
-const { cloudinary } = require('../cloudinary')
+const Campground = require('../models/campground');
+const mbxGeocoding = require('@mapbox/mapbox-sdk/services/geocoding');
+const mapBoxToken = process.env.MAPBOX_TOKEN;
+const geocoder = mbxGeocoding({accessToken:mapBoxToken});
+const { cloudinary } = require('../cloudinary');
 
 module.exports.index = async (req,res)=>{
     const campgrounds = await Campground.find({}); 
-    res.render('campgrounds/index',{campgrounds})
+    res.render('campgrounds/index',{campgrounds});
 }
 module.exports.renderNewForm = (req,res)=>{
-    res.render('campgrounds/new')
+    res.render('campgrounds/new');
 }
 
 module.exports.createCampground =  async (req,res,next)=>{
@@ -17,13 +17,13 @@ module.exports.createCampground =  async (req,res,next)=>{
         query:req.body.campground.location,
         limit:1
     }).send()
-    const campground = new Campground(req.body.campground)
-    campground.geometry = geoData.body.features[0].geometry
-    campground.images = req.files.map(f =>({url:f.path,filename:f.filename}))
-    campground.author = req.user._id
+    const campground = new Campground(req.body.campground);
+    campground.geometry = geoData.body.features[0].geometry;
+    campground.images = req.files.map(f =>({url:f.path,filename:f.filename}));
+    campground.author = req.user._id;
     await campground.save();
-    req.flash('success','Successfully made a new campground')
-    res.redirect(`/campgrounds/${campground._id}`)
+    req.flash('success','Successfully made a new campground');
+    res.redirect(`/campgrounds/${campground._id}`);
 
 }
 
@@ -33,7 +33,7 @@ module.exports.showCampground = async(req,res)=>{
 
     if(!campground){
         req.flash('error','Campground not found')
-         res.redirect('/campgrounds')
+        return res.redirect('/campgrounds')
     }
      res.render('campgrounds/show',{campground})
 
@@ -68,7 +68,12 @@ module.exports.updateCampground = async(req,res)=>{
 
 module.exports.deleteCampground = async (req,res)=>{
     const {id} = req.params
-     await Campground.findByIdAndDelete(id)
+    const campground = await Campground.findById(id);
+     if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'No permission');
+        return res.redirect(`/campgrounds/${id}`);
+        }
+    await Campground.findByIdAndDelete(id)
      req.flash('success','Successfully deleted Campground')
      res.redirect('/campgrounds')
 }
